feat: Use clusters timeline as default
