+++ src/photos/components/AppRoute.jsx
@@ -3,7 +3,6 @@ import { Route, Redirect } from 'react-router'
 
 import Layout from './Layout'
 import Timeline from '../ducks/timeline'
-import TimelineClusters from '../ducks/timeline-clusters'
 import { AlbumsView, AlbumPhotos, PhotosPicker } from '../ducks/albums'
 import PhotosViewer from '../components/PhotosViewer'
 
@@ -18,9 +17,6 @@ const AppRoute = (
         <Route path="edit" component={PhotosPicker} />
         <Route path=":photoId" component={PhotosViewer} />
       </Route>
-    </Route>
-    <Route path="clusters" component={TimelineClusters}>
-      <Route path=":photoId" component={PhotosViewer} />
     </Route>
     <Redirect from="/*" to="photos" />
   </Route>
